Rename Mapper to Properties
Add missing KDocs to Properties
